add : api 에러 메세지 추가, 파일 이름 변경
